Move getAvatarAsPng module
